Refactored SquiggleEditorProps code
